requirements.txt
pandas>=1.0
numpy>=1.18
requests>=2.24
matplotlib>=3.2
python-dotenv>=0.14